src/main/java/com/gildedrose/product/Product.java
package com.gildedrose.product;

public abstract class Product {

    protected String name;
    protected int sellIn;
    protected int quality;

    public Product(String name, int sellIn, int quality) {
        this.name = name;
        this.sellIn = sellIn;
        this.quality = quality;
    }

    public String getName() {
        return name;
    }

    public int getSellIn() {
        return sellIn;
    }

    public int getQuality() {
        return quality;
    }

    public abstract void updateSellIn();

    public abstract void updateQuality();

    @Override
    public String toString() {
        return this.name + ", " + this.sellIn + ", " + this.quality;
    }
}
